+++ src/xrAICore/Navigation/edge_path_inline.h
@@ -35,10 +35,10 @@ TEMPLATE_SPECIALIZATION
 inline void CEdgePathBuilder::get_edge_path(xr_vector<TEdge>& path, Vertex* best, bool reverse_order)
 {
     Vertex *t1 = best, *t2 = best->back();
-    u32 i;
+    size_t i;
     for (i = 1; t2; t1 = t2, t2 = t2->back(), i++)
         ;
-    u32 n = (u32)path.size();
+    size_t n = path.size();
     i--;
     path.resize(n + i);
     t2 = best;
