+++ src/xrCore/_vector4.h
@@ -15,8 +15,8 @@ struct _vector4
 public:
     T x, y, z, w;
 
-    IC T& operator[](int i) { return *((T*)this + i); }
-    IC T& operator[](int i) const { return *((T*)this + i); }
+    IC T& operator[](size_t i) { return *((T*)this + i); }
+    IC T& operator[](size_t i) const { return *((T*)this + i); }
     IC SelfRef set(T _x, T _y, T _z, T _w = 1)
     {
         x = _x;
