+++ src/xrEngine/xr_input.cpp
@@ -472,7 +472,7 @@ void CInput::iRelease(IInputReceiver* p)
     else
     {
         // we are not topmost receiver, so remove the nearest one
-        u32 cnt = cbStack.size();
+        size_t cnt = cbStack.size();
         for (; cnt > 0; --cnt)
             if (cbStack[cnt - 1] == p)
             {
